Improved completionRequiresEngagement, terminating when rewarded odour
Improved method which allows repoking on first odour when completionRequiresEngagement = False to continue sequence without aborting. Terminating and moving onto next odour in any other case.

A method that terminates the trial and makes reward available if the odour is rewarded which should work for either case of completionRequiresEngagement.

Insert a timeout period for completionRequiresEngagement=True using the same variable as interTrialIntervalUnsuccessfulTrial.